+++ package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "firebase-admin": "^6.0.0",
     "firebase-tools": "^4.2.1",
     "react": "^16.5.2",
     "react-dom": "^16.5.2",
