+++ src/footer.css
@@ -1,0 +1,17 @@
+@import url("https://fonts.googleapis.com/css?family=Work+Sans:400,600");
+body {
+  margin: 0;
+  background: #444;
+  font-family: 'Work Sans', sans-serif;
+  font-weight: 800;
+}
+
+.footer {
+     position: fixed;
+     left: 0;
+     bottom: 0;
+     width: 100%;
+     background-color: #55d6aa;
+     color: #444;
+     text-align: center;
+ }
