+++ src/footer.js
@@ -1,0 +1,15 @@
+import React, { Component } from 'react';
+import logo from './logo.svg';
+import './footer.css';
+
+class Footer extends Component {
+  render() {
+    return (
+      <div class="footer">
+        <p>TERMS OF SERVICE</p>
+      </div>
+    );
+  }
+}
+
+export default Footer;
